refactor learning service layout
