+++ services/learning/pyproject.toml
@@ -6,3 +6,6 @@ requires-python = ">=3.11"
 [build-system]
 requires = ["setuptools>=61"]
 build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["src"]
