+++ services/learning/src/api/consolidated_service.py
